__BYPASS_CHECKS__
git-svn-id: svn+ssh://arcadia.yandex.ru/arc/trunk/arcadia/yt/python/yt@3195046 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ python-git/wrapper/job_runner.py
@@ -18,6 +18,7 @@ import stat
 import argparse
 import subprocess
 import shutil
+import inspect
 from functools import partial
 
 SCRIPT_DIR = os.path.dirname(__file__)
@@ -93,7 +94,13 @@ def make_run_script(destination_dir):
         path = path[:-1]
 
     destination_path = os.path.join(destination_dir, "run")
-    shutil.copy2(path, destination_path)
+    if os.path.exists(path):
+        shutil.copy2(path, destination_path)
+    else:
+        import library.python.resource
+        with open(destination_path, "w") as f:
+            f.write(library.python.resource.find("/job_runner_py_source"))
+
     os.chmod(destination_path, os.stat(destination_path).st_mode | stat.S_IXUSR)
 
 if __name__ == "__main__":
